split xtarget
