+++ CMakeLists.txt
@@ -105,6 +105,7 @@ set(XWIDGETS_HEADERS
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xstyle.hpp
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xradiobuttons.hpp
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xtab.hpp
+    ${XWIDGETS_INCLUDE_DIR}/xwidgets/xtarget.hpp
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xtext.hpp
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xtextarea.hpp
     ${XWIDGETS_INCLUDE_DIR}/xwidgets/xtogglebutton.hpp
@@ -140,6 +141,7 @@ set(XWIDGETS_SOURCES
     ${XWIDGETS_SOURCE_DIR}/xslider.cpp
     ${XWIDGETS_SOURCE_DIR}/xselectionslider.cpp
     ${XWIDGETS_SOURCE_DIR}/xtab.cpp
+    ${XWIDGETS_SOURCE_DIR}/xtarget.cpp
     ${XWIDGETS_SOURCE_DIR}/xtext.cpp
     ${XWIDGETS_SOURCE_DIR}/xtextarea.cpp
     ${XWIDGETS_SOURCE_DIR}/xtogglebutton.cpp
