+++ include/xwidgets/xtarget.hpp
@@ -1,0 +1,21 @@
+/***************************************************************************
+* Copyright (c) 2017, Sylvain Corlay and Johan Mabille                     *
+*                                                                          *
+* Distributed under the terms of the BSD 3-Clause License.                 *
+*                                                                          *
+* The full license is in the file LICENSE, distributed with this software. *
+****************************************************************************/
+
+#ifndef XWIDGETS_TARGET_HPP
+#define XWIDGETS_TARGET_HPP
+
+#include "xeus/xcomm.hpp"
+
+#include "xwidgets_config.hpp"
+
+namespace xw
+{
+    XWIDGETS_API xeus::xtarget* get_widget_target();
+}
+
+#endif
